refactor(core): code refactor
